+++ frontend/src/components/PrivateRoute.jsx
@@ -1,22 +1,12 @@
-import React, { useEffect } from 'react';
-import { Route, Redirect, useHistory } from 'react-router-dom';
+import React from 'react';
+import { Route, Redirect } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import Cookies from 'js.cookie';
 import { JWT_REFRESH_TOKEN } from '../constants/cookies';
 
 const PrivateRoute = ({ component: Component, ...rest }) => {
-  const history = useHistory();
-
   const refreshToken = Cookies.get(JWT_REFRESH_TOKEN);
   const currentUser = useSelector((state) => state.userReducer.currentUser);
-  const currentUserRequest = useSelector(
-    (state) => state.loadingReducer['user/GET_CURRENT_USER']
-  );
-
-  useEffect(() => {
-    if (currentUserRequest !== 'FAILURE') return;
-    history.push('/login');
-  }, [currentUserRequest]);
 
   return (
     <Route
