+++ frontend/src/components/posts/NewPost.jsx
@@ -10,7 +10,6 @@ const NewPostWrapper = styled.div`
     box-shadow: 0 2px 2px rgba(154, 160, 185, 0.05),
       0 5px 5px rgba(166, 173, 201, 0.1);
   }
-  width: 650px;
   margin: 0 auto;
   border: 1px solid #eee;
   padding: 10px;
